pyproject.toml
[project]
name = "gptsh-cli"
version = "0.4.1"
requires-python = ">=3.10"
dependencies = [
  "litellm",
  "click",
  "pyyaml",
  "rich",
  "httpx",
  "python-dotenv",
  "mcp>=1.15.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  # linters/formatters (ruff is preferred primary linter)
  "ruff",
]

[project.scripts]
gptsh = "gptsh.cli.entrypoint:main"

[tool.ruff]
line-length = 100
target-version = "py310"
respect-gitignore = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "B",   # flake8-bugbear
  "I",   # isort
  "W",   # warnings
]
ignore = [
  "E203",  # whitespace before ':' (black compatibility)
  "E501",  # Line too long
]
exclude = [
  ".venv",
  ".pytest_cache",
  ".serena",
  ".gptsh",
  "gptsh_cli.egg-info",
]

[tool.ruff.lint.isort]
known-first-party = ["gptsh"]
combine-as-imports = true
